[reactive-element] Consolidate decorator implementations (#4156)
