+++ packages/reactive-element/src/decorators.ts
@@ -8,7 +8,6 @@
 // decorators as implemented by TypeScript 5.2 and Babel, and with TypeScript's
 // experimentalDecorators option.
 
-export * from './decorators/base.js';
 export * from './decorators/custom-element.js';
 export * from './decorators/property.js';
 export * from './decorators/state.js';
