--- packages/reactive-element/src/legacy-decorators/query-async.ts
@@ -1,93 +1,0 @@
-/**
- * @license
- * Copyright 2017 Google LLC
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-/*
- * IMPORTANT: For compatibility with tsickle and the Closure JS compiler, all
- * property decorators (but not class decorators) in this file that have
- * an @ExportDecoratedItems annotation must be defined as a regular function,
- * not an arrow function.
- */
-
-import type {ReactiveElement} from '../reactive-element.js';
-import {Interface} from './base.js';
-import {queryAsync as standardQueryAsync} from '../std-decorators/query-async.js';
-
-export type QueryAsyncDecorator = {
-  // legacy
-  (
-    proto: Interface<ReactiveElement>,
-    name: PropertyKey
-    // Note TypeScript requires the return type to be `void|any`
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  ): void | any;
-
-  // standard
-  <C extends Interface<ReactiveElement>, V extends Promise<Element | null>>(
-    value: ClassAccessorDecoratorTarget<C, V>,
-    context: ClassAccessorDecoratorContext<C, V>
-  ): void;
-};
-
-// Note, in the future, we may extend this decorator to support the use case
-// where the queried element may need to do work to become ready to interact
-// with (e.g. load some implementation code). If so, we might elect to
-// add a second argument defining a function that can be run to make the
-// queried element loaded/updated/ready.
-/**
- * A property decorator that converts a class property into a getter that
- * returns a promise that resolves to the result of a querySelector on the
- * element's renderRoot done after the element's `updateComplete` promise
- * resolves. When the queried property may change with element state, this
- * decorator can be used instead of requiring users to await the
- * `updateComplete` before accessing the property.
- *
- * @param selector A DOMString containing one or more selectors to match.
- *
- * See: https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelector
- *
- * ```ts
- * class MyElement {
- *   @queryAsync('#first')
- *   first: Promise<HTMLDivElement>;
- *
- *   render() {
- *     return html`
- *       <div id="first"></div>
- *       <div id="second"></div>
- *     `;
- *   }
- * }
- *
- * // external usage
- * async doSomethingWithFirst() {
- *  (await aMyElement.first).doSomething();
- * }
- * ```
- * @category Decorator
- */
-export function queryAsync(selector: string) {
-  return (<C extends Interface<ReactiveElement>, V extends Promise<Element>>(
-    protoOrTarget: ClassAccessorDecoratorTarget<C, V>,
-    nameOrContext: PropertyKey | ClassAccessorDecoratorContext<C, V>
-  ) => {
-    if (typeof nameOrContext === 'object') {
-      return standardQueryAsync(selector)(
-        protoOrTarget,
-        nameOrContext as ClassAccessorDecoratorContext<C, V>
-      );
-    } else {
-      Object.defineProperty(protoOrTarget, nameOrContext as PropertyKey, {
-        async get(this: ReactiveElement) {
-          await this.updateComplete;
-          return this.renderRoot?.querySelector(selector);
-        },
-        enumerable: true,
-        configurable: true,
-      });
-      return;
-    }
-  }) as QueryAsyncDecorator;
-}
