--- packages/reactive-element/src/std-decorators/custom-element.ts
@@ -1,39 +1,0 @@
-/**
- * @license
- * Copyright 2023 Google LLC
- * SPDX-License-Identifier: BSD-3-Clause
- */
-
-/*
- * IMPORTANT: For compatibility with tsickle and the Closure JS compiler, all
- * property decorators (but not class decorators) in this file that have
- * an @ExportDecoratedItems annotation must be defined as a regular function,
- * not an arrow function.
- */
-
-import type {Constructor} from '../legacy-decorators/base.js';
-
-/**
- * Class decorator factory that defines the decorated class as a custom element.
- *
- * ```js
- * @customElement('my-element')
- * class MyElement extends LitElement {
- *   render() {
- *     return html``;
- *   }
- * }
- * ```
- * @category Decorator
- * @param tagName The tag name of the custom element to define.
- */
-export const customElement =
-  (tagName: string) =>
-  (
-    _value: unknown,
-    {addInitializer}: ClassDecoratorContext<Constructor<HTMLElement>>
-  ) => {
-    addInitializer(function () {
-      customElements.define(tagName, this);
-    });
-  };
